Moved PluginUpdatesDialog into picard.ui
